requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
asgiref==3.7.2
    # via django
azure-core==1.28.0
    # via azure-storage-blob
azure-identity==1.14.0
    # via azure-storage-blob
azure-storage-blob==12.17.0
    # via django-storages
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
cryptography==41.0.3
    # via azure-storage-blob
dj-database-url==2.0.0
    # via -r src/requirements/requirements.in
django==4.1.10
    # via
    #   -r src/requirements/requirements.in
    #   dj-database-url
    #   django-storages
django-storages[azure]==1.13.2
    # via -r src/requirements/requirements.in

gunicorn==21.2.0
    # via -r src/requirements/requirements.in  
idna==3.4
    # via requests
isodate==0.6.1
    # via azure-storage-blob
pillow==10.0.0
    # via -r src/requirements/requirements.in
psycopg2-binary==2.9.6
    # via -r src/requirements/requirements.in
pycparser==2.21
    # via cffi
python-decouple==3.8
    # via -r src/requirements/requirements.in
requests==2.31.0
    # via
    #   azure-core
    #   stripe
six==1.16.0
    # via
    #   azure-core
    #   isodate
sqlparse==0.4.4
    # via django
stripe==5.5.0
    # via -r src/requirements/requirements.in
typing-extensions==4.7.1
    # via
    #   asgiref
    #   azure-core
    #   azure-storage-blob
    #   dj-database-url
urllib3==2.0.4
    # via requests

whitenoise==6.5.0
